+++ src/main/java/org/bob/siungongsi/repository/GongsiRepository.java
@@ -1,5 +1,6 @@
-package org.bob.siungongsi.domain;
+package org.bob.siungongsi.repository;
 
+import org.bob.siungongsi.domain.GongsiEntity;
 import org.springframework.data.repository.CrudRepository;
 
 public interface GongsiRepository extends CrudRepository<GongsiEntity, String> {}
