+++ src/main/java/org/bob/siungongsi/repository/TodayProcessedGongsiRepository.java
@@ -1,5 +1,6 @@
-package org.bob.siungongsi.domain;
+package org.bob.siungongsi.repository;
 
+import org.bob.siungongsi.domain.TodayProcessedGongsiEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TodayProcessedGongsiRepository
