+++ src/main/java/org/bob/siungongsi/service/MessageListener.java
@@ -11,12 +11,12 @@ import java.util.concurrent.ExecutionException;
 import org.bob.siungongsi.client.OpenDartReader;
 import org.bob.siungongsi.config.SqsProperties;
 import org.bob.siungongsi.domain.CompanyEntity;
-import org.bob.siungongsi.domain.CompanyRepository;
 import org.bob.siungongsi.domain.GongsiEntity;
-import org.bob.siungongsi.domain.GongsiRepository;
 import org.bob.siungongsi.domain.ProcessingFailedGongsiEntity;
-import org.bob.siungongsi.domain.ProcessingFailedGongsiRepository;
 import org.bob.siungongsi.event.GongsiMessage;
+import org.bob.siungongsi.repository.CompanyRepository;
+import org.bob.siungongsi.repository.GongsiRepository;
+import org.bob.siungongsi.repository.ProcessingFailedGongsiRepository;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.ApplicationListener;
 import org.springframework.context.event.ContextClosedEvent;
